Technical/Refactor functions and methods scope
* Extracted private packet helpers, tests
* Extracted private packet functions to ConfigurationAttr methods, tests
* Simplified NewConfiguration
* Moved all consts to independent file
