+++ helpers_test.go
@@ -1,119 +1,70 @@
 package truemail
 
 import (
-	"fmt"
-	"net"
-	"strconv"
+	"regexp"
+	"testing"
 
-	"github.com/brianvoe/gofakeit/v6"
-	"github.com/foxcpp/go-mockdns"
+	"github.com/stretchr/testify/assert"
 )
 
-func randomEmail() string {
-	gofakeit.Seed(0)
-	return gofakeit.Email()
+func TestIsIncluded(t *testing.T) {
+	var item string
+
+	t.Run("item found in slice", func(t *testing.T) {
+		assert.True(t, isIncluded([]string{item}, item))
+	})
+
+	t.Run("item not found in slice", func(t *testing.T) {
+		assert.False(t, isIncluded([]string{}, item))
+	})
 }
 
-func randomDomain() string {
-	gofakeit.Seed(0)
-	return gofakeit.DomainName()
+func TestIsIntersected(t *testing.T) {
+	var item string
+
+	t.Run("item from target slice found in base slice", func(t *testing.T) {
+		assert.True(t, isIntersected([]string{item}, []string{item}))
+	})
+
+	t.Run("item from target slice not found in base slice", func(t *testing.T) {
+		assert.False(t, isIntersected([]string{}, []string{item}))
+	})
 }
 
-func pairRandomEmailDomain() (string, string) {
-	gofakeit.Seed(0)
-	domain := randomDomain()
-	email := gofakeit.Username() + "@" + domain
-	return email, domain
+func TestNewRegex(t *testing.T) {
+	t.Run("valid regex pattern", func(t *testing.T) {
+		regexPattern := EmptyString
+		actualRegex, err := newRegex(regexPattern)
+		expectedRegex, _ := regexp.Compile(regexPattern)
+
+		assert.Equal(t, expectedRegex, actualRegex)
+		assert.NoError(t, err)
+	})
+
+	t.Run("invalid regex pattern", func(t *testing.T) {
+		actualRegex, err := newRegex(`\K`)
+
+		assert.Nil(t, actualRegex)
+		assert.Error(t, err)
+	})
 }
 
-func randomIpAddress() string {
-	gofakeit.Seed(0)
-	return gofakeit.IPv4Address()
+func TestMatchRegex(t *testing.T) {
+	t.Run("valid regex pattern, matched string", func(t *testing.T) {
+		assert.True(t, matchRegex(EmptyString, EmptyString))
+	})
+
+	t.Run("valid regex pattern, not matched string", func(t *testing.T) {
+		assert.False(t, matchRegex("42", `\D+`))
+	})
+
+	t.Run("invalid regex pattern", func(t *testing.T) {
+		assert.False(t, matchRegex(EmptyString, `\K`))
+	})
 }
 
-func randomIp6Address() string {
-	gofakeit.Seed(0)
-	return gofakeit.IPv6Address()
-}
-
-func randomPortNumber() int {
-	gofakeit.Seed(0)
-	return gofakeit.Number(1, 65535)
-}
-
-func randomDnsServer() string {
-	gofakeit.Seed(0)
-	return randomIpAddress() + ":" + strconv.Itoa(randomPortNumber())
-}
-
-func createConfiguration() *configuration {
-	configuration, _ := NewConfiguration(ConfigurationAttr{verifierEmail: randomEmail()})
-	return configuration
-}
-
-func createValidatorResult(email string, configuration *configuration, options ...string) *validatorResult {
-	validationType, _ := variadicValidationType(options)
-	validatorResult := &validatorResult{Email: email, Configuration: configuration, ValidationType: validationType}
-	return validatorResult
-}
-
-func createSuccessfulValidatorResult(email string, configuration *configuration) *validatorResult {
-	return &validatorResult{Email: email, Configuration: configuration, Success: true}
-}
-
-func randomValidationType() string {
-	gofakeit.Seed(0)
-	availableValidationTypes := []string{ValidationTypeRegex, ValidationTypeMx, ValidationTypeSMTP}
-	index := gofakeit.Number(0, len(availableValidationTypes)-1)
-	return availableValidationTypes[index]
-}
-
-func createValidator(email string, configuration *configuration, options ...string) *validator {
-	validationType, _ := variadicValidationType(options)
-	return newValidator(email, validationType, configuration)
-}
-
-func usedValidationsByType(validationType string) []string {
-	return map[string][]string{
-		ValidationTypeRegex:       {ValidationTypeRegex},
-		ValidationTypeMx:          {ValidationTypeRegex, ValidationTypeMx},
-		ValidationTypeMxBlacklist: {ValidationTypeRegex, ValidationTypeMx, ValidationTypeMxBlacklist},
-		ValidationTypeSMTP:        {ValidationTypeRegex, ValidationTypeMx, ValidationTypeMxBlacklist, ValidationTypeSMTP},
-	}[validationType]
-}
-
-func runDomainListMatchValidation(email string, configuration *configuration, options ...string) *validatorResult {
-	validator := createValidator(email, configuration, options...)
-	validatorResult := validator.result
-	return validator.domainListMatch.check(validatorResult)
-}
-
-func doPassedFromDomainListMatch(validatorResult *validatorResult) {
-	validatorResult.Success, validatorResult.isPassFromDomainListMatch = true, true
-}
-
-func failedValidatorResult() *validatorResult {
-	return new(validatorResult)
-}
-
-// Returns dnsResolver with mocked DNS records
-func createDnsResolver(dnsRecords map[string]mockdns.Zone) *dnsResolver {
-	return &dnsResolver{gateway: &mockdns.Resolver{Zones: dnsRecords}}
-}
-
-func createDnsResolverWithEpmtyRecords() *dnsResolver {
-	return createDnsResolver(map[string]mockdns.Zone{})
-}
-
-func dnsErrorMessage(hostname string) string {
-	return fmt.Sprintf("lookup %s on 127.0.0.1:53: no such host", hostname)
-}
-
-func runMockDnsServer(dnsRecords map[string]mockdns.Zone) string { // TODO: how to remove DNS request stdout dig log?
-	srv, _ := mockdns.NewServer(dnsRecords, false)
-	runningMockServerAddress := srv.LocalAddr().String()
-	defer srv.Close()
-	srv.PatchNet(net.DefaultResolver)
-	defer mockdns.UnpatchNet(net.DefaultResolver)
-	return runningMockServerAddress
+func TestAvailableValidationTypes(t *testing.T) {
+	t.Run("slice of available validation types", func(t *testing.T) {
+		assert.Equal(t, []string{"regex", "mx", "mx_blacklist", "smtp"}, availableValidationTypes())
+	})
 }
